% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_download.R
\name{cached_download}
\alias{cached_download}
\title{Download a file and cache it locally}
\usage{
cached_download(url, cache_dir = NULL)
}
\arguments{
\item{url}{url to download}

\item{cache_dir}{directory to which file should be saved. If not
provided, it will be saved in the package's directory.}
}
\value{
path to the downloaded file
}
\description{
Download a file and cache it locally
}
