% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plants.R
\name{read_plants}
\alias{read_plants}
\title{Read plant data from form EIA-860}
\usage{
read_plants(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{directory in which to cache downloaded files; will be cached
in package internal directory by default}
}
\value{
data frame containing plant data
}
\description{
For each plant, compute the dominant fuel type.
}
