% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_year.R
\name{read_year}
\alias{read_year}
\title{Read generation data for a single year}
\usage{
read_year(year, cache_dir = NULL)
}
\arguments{
\item{year}{year to read}

\item{cache_dir}{directory in which to cache downloaded files; will be cached
in package internal directory by default}
}
\value{
data frame containing generation data
}
\description{
Read generation data for a single year
}
