% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_for_year.R
\name{url_for_year}
\alias{url_for_year}
\title{Return the download URL for a given year's EIA-923 data}
\usage{
url_for_year(year)
}
\arguments{
\item{year}{year of data}
}
\value{
url to zipfile
}
\description{
Return the download URL for a given year's EIA-923 data
}
