% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_filename.R
\name{xl_filename}
\alias{xl_filename}
\title{Returns the Excel file name used for a given year's data}
\usage{
xl_filename(year)
}
\arguments{
\item{year}{year}
}
\value{
filename
}
\description{
Returns the Excel file name used for a given year's data
}
